pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.14",
]
